fix: 🧩 聚合webpack命令
